more_singly_linked_lists_tasks

+++ 0x13-more_singly_linked_lists/9-insert_nodeint.c
@@ -29,7 +29,7 @@ listint_t *insert_nodeint_at_index(listint_t **head, unsigned int idx, int n)
 
 	for (i = 0; *head != NULL; i++)
 	{
-		if (i + 1 == (int) idx)
+		if (i + 1 == idx)
 		{
 			new->next = (*head)->next;
